+++ Polly.Contrib.AzureFunctions/CircuitBreaker/CircuitState.cs
@@ -2,8 +2,8 @@
 {
     public enum CircuitState
     {
-        Closed,
-        Open,
-        HalfOpen,
+        Closed = 0,
+        Open = 1,
+        HalfOpen = 2,
     }
 }